chrome mods, study_base_classes into loader.
For debugging old studies, it is better to load
the required stuff as globals/pseudomodules.  This is probably
the wrong solution.

Right solution:  flag for 'old' that loads the oldsdk/ env?

Sharp edge:  if study_base_classes moves in testpilotweb,
it has to get updated here (which is the reason it's tpweb at
all, so that we can sneak code changes in easily!)